refactor(eventhandlers): specify WebhookEvent content expectations in its type (#430)
